+++ SmartVideo.framework/Modules/module.modulemap
@@ -1,11 +1,4 @@
-framework module SmartVideo {
-  umbrella header "SmartVideo.h"
-
-  export *
-  module * { export * }
-}
-
-module SmartVideo.Swift {
-    header "SmartVideo-Swift.h"
-    requires objc
+framework module SmartVideoSDK {
+  header "SmartVideoSDK-Swift.h"
+  requires objc
 }
